requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-google-native>=0.7.0
google-cloud-bigquery>=3.6.0
